Parser changes Compiler Changes
Now the parser is ready to understand expressions but the compiler still does not transform them.

The compiler now transforms integers and floats that are not type compatible via a cast, similar to u8 to u64 in Rust.

The parser now checks whether the cast is possible or not. For example, a u64 does not fit, it is a u8 variable.