New Compiler Organization and Important Changes in the API
The general structure of the compiler was changed to one that was more functional and enjoyable for the programmer.

Memory Reference Counting implemented to free Heap pointers

Now the compiler uses the Vec<?> vector api to allocate, reallocate string dynamically in memory.

Increasingly the Vector API is becoming more stable for use.

In the command line it has changed that now, through --backend-args, the path to the object file that contains the api vector is specified.

There are now methods for types i8, i16, i32 and i64 to add an element to the dynamic array.