String Changes, Variables Changes
Now the strings compile successfully, being mutated, or referenced.

In the case of taking a reference from another string, the referenced string is completely cloned to the new variable, behind it it points to the same memory address so it ends up using only one @Vec.destroy(ptr) & @free(ptr) .

Additionally, there is now deeper experimental type checking in the language.